Move createMQConnector and getMQConnectorClass to MQconnector.py
